fix: clear buttons before sending if one page async iterator source
Moves the call to clear into send_initial_message instead of start so that with async iterator page source, it will not send the buttons if there is only one page.

Menu.clear() was fixed to only clear buttons and not all items, and additionally not call message.edit if message is None.

For consistency and performance, ButtonMenu.disable() and ButtonMenu.enable() similarly will not edit the message if the view has not changed.